module5-task2 | Перенос констант в папку utils
