+++ src/components/catalog/offer-card/offer-card.tsx
@@ -1,6 +1,6 @@
 import {Link} from 'react-router-dom';
 import {Offer} from '../../../types/offer';
-import {AppRoute} from '../../../const';
+import {AppRoute} from '../../../utils/const';
 import {clsx} from 'clsx';
 import {getRatingWidth} from '../../../utils/common';
 import OfferBookmark from '../offer-bookmark/offer-bookmark';
