+++ src/components/catalog/offer-review-form/const.ts
@@ -1,0 +1,7 @@
+export const RATING = [
+  {value: 5, title: 'perfect'},
+  {value: 4, title: 'good'},
+  {value: 3, title: 'not bad'},
+  {value: 2, title: 'badly'},
+  {value: 1, title: 'terribly'},
+];
