+++ src/components/common/app/app.tsx
@@ -3,7 +3,7 @@ import MainPage from '../../../pages/main-page/main-page';
 import LoginPage from '../../../pages/login-page/login-page';
 import FavoritesPage from '../../../pages/favorites-page/favorites-page';
 import NotFoundPage from '../../../pages/not-found-page/not-found-page';
-import {AppRoute} from '../../../const';
+import {AppRoute} from '../../../utils/const';
 import OfferPage from '../../../pages/offer-page/offer-page';
 import {Offer} from '../../../types/offer';
 import ProtectedRoute from '../protected-route/protected-route';
