+++ src/components/common/header/header.tsx
@@ -1,5 +1,5 @@
 import {Link} from 'react-router-dom';
-import {AppRoute} from '../../../const';
+import {AppRoute} from '../../../utils/const';
 import HeaderLogo from '../header-logo/header-logo';
 
 export default function Header(): JSX.Element {
