+++ src/components/common/protected-route/protected-route.tsx
@@ -1,5 +1,5 @@
 import {Navigate, useLocation} from 'react-router-dom';
-import {AppRoute, AuthStatus} from '../../../const';
+import {AppRoute, AuthStatus} from '../../../utils/const';
 import {setAuthStatus} from '../../../utils/common';
 import {Location} from 'react-router-dom';
