+++ src/hooks/use-map-leaflet-markers/const.ts
@@ -1,0 +1,4 @@
+import URL_MARKER_DEFAULT from './pin.svg';
+import URL_MARKER_CURRENT from './main-pin.svg';
+
+export {URL_MARKER_DEFAULT, URL_MARKER_CURRENT};
