+++ src/pages/not-found-page/not-found-page.tsx
@@ -2,7 +2,7 @@ import Container from '../../components/common/container/container';
 import Header from '../../components/common/header/header';
 import Footer from '../../components/common/footer/footer';
 import {Link, useLocation} from 'react-router-dom';
-import {AppRoute} from '../../const';
+import {AppRoute} from '../../utils/const';
 import MainContainer from '../../components/common/main-container/main-container';
 import {Helmet} from 'react-helmet-async';
