+++ src/utils/const.ts
@@ -1,4 +1,4 @@
-import {City} from './types/city';
+import {City} from '../types/city';
 
 export enum AppRoute {
   Root = '/',
@@ -27,16 +27,3 @@ export const DEFAULT_CITY: City = {
     zoom: 12
   }
 };
-
-export const URL_MARKER_DEFAULT = 'https://assets.htmlacademy.ru/content/intensive/javascript-1/demo/interactive-map/pin.svg';
-
-export const URL_MARKER_CURRENT = 'https://assets.htmlacademy.ru/content/intensive/javascript-1/demo/interactive-map/main-pin.svg';
-
-
-export const RATING = [
-  {value: 5, title: 'perfect'},
-  {value: 4, title: 'good'},
-  {value: 3, title: 'not bad'},
-  {value: 2, title: 'badly'},
-  {value: 1, title: 'terribly'},
-];
